Fix delta time
Delta time used to be a const value of 60 frames per second, but now it is correctly updated based on your computers performance. Note that even though delta time is now correctly being updated, it remains unused due to the fact that applying it would break the implicit 10 size grid-layout of the game. This value of 10 isn't specified anywhere 'exactly', it's just 'coincidentally' used across the code base. Therefor, a grid size has to be specified (likely by the game) that all other values are controlled by. Only then may delta time become of use.