Reverse magic value declaration
Magic value for the game width and height were declared inside Game.h and Game.cpp, these values being the width, height, and title of the window. The magic values are now declared inside of main and then passed to Game so it is more clear where these values are coming from. Note that this way of declaring the width and height will soon be outdated. In later stages, the game will rely on a gridsize to calculate its final size.