Add a bad commit for learning purposes
Originally, I was thinking of making everything work via a grid. As I went on, somewhere I decided it was a good idea to remove the Player and the apple from the game, so that GameObjects would be decoupled from it. They could then add themselves or remove themselves from the game process, but for this I needed a very rudementary component system, so I made IUpdateable and IRenderable...

Long story short, this is what it looks like when a commit goes too far. This isn't refactoring anymore: this is starting from scratch. I'm pushing this commit message for the purpose that Ulf, my teacher for this course, is able to see that I can learn from my mistakes. I will be reverting this commit immediately after pushing.