Use approximatelyEquals comparison for Vector2
Vector2 is build on floating point numbers. Because of floating point imprecision, comparing these may result in undesired behaviour of comparing two vectors that are seemingly equal resulting in a null comparison.

It has not escaped me that floating point comparison is a complex subject with a lot of nuance. Many articles about its complexity can be found: the two below are just the ones I've read. However, we may accept that, for our purposes, this implementation will do.

https://stackoverflow.com/questions/17333/what-is-the-most-effective-way-for-float-and-double-comparison
https://www.boost.org/doc/libs/1_67_0/libs/math/doc/html/math_toolkit/float_comparison.html