Make the Player frame-independant
The (stupid) 20 fps cap has been lifted for this game and now runs at a smooth whatever-your-processor-will-allow fps! This approach proved however to have multiple difficulties regarding the snake staying inside of the grid, since this was hard coded in the movement (it was pretty bad).

The input therefor always snaps the headPosition on the grid. This way it may never escape the grid (and thus never cause any undesired collision between body parts it is 'slightly' touching). The bodyPositions, while tricky to figure out, behave in a similar manner. The bodyPositions themselves will always snap to the grid and never move by themselves. Instead, once the snake moves one grid-space away, the last bodyPosition is removed while a new one is placed at the head of the snake. We also compensate for any jumps the snake makes inside of UpdateInput().

Currently the tail-end of the snake looks stutery, however this may be easily fixed by providing a "tailPosition" that updates smoothly over time. In this position, the snake can give the impression of smooth movement while staying as optimized as possible.