Add a tail position to the player
The tail position is updated to give the impression of smooth movement. The implementation suffers from not staying perfectly on the grid at all times. However, rather than working this out, the Player class will be deprecated in favor of a new class build up from scratch to provide a cleaner solution.