Replace type Player with type Snake
The new Snake class, previously Player class, puts its Head and Tail inside the bodyParts deque and ensures the size of the deque may never shrink below 2. This allows for some optimizations, as well as a better coding workflow that repeats less lines of code.

It currently provides smooth movement (tail not included) that, just like the Player class, makes use of the deque to simply put its last body part at the front. What's interesting to note is that in the Snake class, putting a new body part at the front actually 'makes' it the head, while the previous head turns into "the second body part".