Make Update Body independant of distance
The previous method for updating the body would update the body every time the head was registered on "the grid" (the grid is made up of the Snake's size). However, if framerate is low or speed is high, these updates may accidentally be skipped.

The new method does not check if the head is perfectly on the grid, but rather checks the distance between the head and the second body part. Based on that distance, an amount of body parts is then removed from the back and added in a row to the front. Even with framerates of 4 (yes I actually tested that) the body had no problem updating correctly.

Additionally, the Snake should have no problem warping to a different place as long as UpdateHead() and UpdateBody() are called immediately after.