Add UpdateTail method
UpdateTail will update the tail, and we don't need an extra variable to do it! The snake always consists of a head and a tail and the tail is always the last body part of the snake. We use that knowledge to our advantage. By first calculating the distance from the snake to the second body part, we can take our size minus that distance to find out how far the tail should be from the second to last body part. The trick looks a bit weird in code, but the results have never been more smooth!