Provide public API for common functions
SnapPositionToSize() has been added for users to snap the Snake position to a place on the grid if they so desire. This might be useful if they want to stop the snake is place somewhere and would like it to snap to the grid during that time (for example through a pause feature).

ChangeDirection(Vector2 direction) can be used to change the direction of the snake. It will not allow this if the provided direction is the same as that of what the snake already has or the inverse of that. If the direction is changed, measures are taken to make sure the snake stays unbroken as well as snappy.

ProcessInput has been refactored to simple make a function call to ChangeDirection with a parameter based on the provided input.